Major refactors and bug fixes
