+++ model/members.js
@@ -1,9 +1,6 @@
 'use strict';
 
 const mongoose = require('mongoose');
-var config = require('../config.js');
-// mongodb connection
-const DATABASE_URL = process.env.MONGODB_URI || config.database.url;
 
 //define a schema
 var Schema = mongoose.Schema;
@@ -69,63 +66,6 @@ var getMembersWithSkills = function (error, members) {
     console.log(members);
 };
 
-
-
-/**
- * Initialize Mongoose adapter for MongoDB. If connection fails on first attempt, no further attempts to connect 
- * is made. This is a fail fast strategy. If connected to DB then web server is started.
- * If connection fails after initial connection then unlimited reconnect attempts will be made. Web server will be 
- * shutdown until successful connection is established
- */
-try {
-    let options = {
-        "autoReconnect": true,
-        "reconnectTries": Number.MAX_VALUE,
-        "reconnectInterval": 500
-    };
-    mongoose.Promise = global.Promise;
-    mongoose.connect(DATABASE_URL, options).then(function () {
-        console.info(`Mongoose connection successfully created`);
-    }).catch(function (error) {
-        console.error(`Error on connection: ${error}`);
-    });
-    var db = mongoose.connection;
-
-    // When successfully connected
-    mongoose.connection.on('connected', function () {
-        /**
-         * Start the webserver when connection is established
-         */
-        require('../server.js').startup();
-        console.log('Mongoose default connection open to ' + DATABASE_URL);
-    });
-
-    // If the connection throws an error
-    mongoose.connection.on('error', function (err) {
-        console.log('Mongoose default connection error: ' + err);
-    });
-
-    // When the connection is disconnected
-    mongoose.connection.on('disconnected', function () {
-        /**
-         * Shutdown web server if the DB is unavilable
-         */
-        require('../server.js').shutdown();
-        console.log('Mongoose default connection disconnected');
-
-    });
-    // If the Node process ends, close the Mongoose connection 
-    process.on('SIGINT', function () {
-        mongoose.connection.close(function () {
-            console.log('Mongoose default connection disconnected through app termination');
-            process.exit(0);
-        });
-    });
-
-} catch (err) {
-    console.log('Catastrophic error in connecting to MongoDb: ' + err);
-}
-
 module.exports.list = function () {
     return members.find(listMembers).exec();
 };
@@ -160,7 +100,7 @@ module.exports.findMembersBySkills = function (skills) {
  * @param {*}      cb            callback
  */
 module.exports.addMember = function (memberProfile, cb) {
-    cb = typeof (cb) === 'function' ? cb : function () {};
+    cb = typeof (cb) === 'function' ? cb : function () { };
     let newMember = new members(memberProfile);
     newMember.save(function (err, doc) {
         console.debug(doc);
@@ -176,7 +116,7 @@ module.exports.addMember = function (memberProfile, cb) {
  */
 module.exports.updateMember = function (memberProfile, memberId, cb) {
     //check whether the callback is a callable function
-    cb = typeof (cb) === 'function' ? cb : function () {};
+    cb = typeof (cb) === 'function' ? cb : function () { };
 
     //build your update object with modifiedTS updated
     let keyVal = {};
@@ -196,10 +136,10 @@ module.exports.updateMember = function (memberProfile, memberId, cb) {
     members.findByIdAndUpdate(memberId, {
         $set: keyVal
     }, {
-        new: true
-    }, function (error, updatedMember) {
-        cb(error, updatedMember);
-    });
+            new: true
+        }, function (error, updatedMember) {
+            cb(error, updatedMember);
+        });
 };
 
 /**
@@ -208,8 +148,8 @@ module.exports.updateMember = function (memberProfile, memberId, cb) {
  * @param {*} cb Callback function 
  */
 module.exports.removeMember = function (id, cb) {
-    cb = typeof (cb) === 'function' ? cb : function () {};
-    members.findByIdAndRemove(id, function(err, removedMember){
-        cb(err,removedMember);
+    cb = typeof (cb) === 'function' ? cb : function () { };
+    members.findByIdAndRemove(id, function (err, removedMember) {
+        cb(err, removedMember);
     });
 };
